# api/api.yaml
...
openapi: "3.0.0"
info:
  title: Quiet Protocol API
  version: "1.0.0"
  description: |
    REST API for desktop/mobile clients.
    Designed for polling (100ms intervals when foreground).
    All endpoints scoped to the peer running the API server.
servers:
  - url: http://localhost:8080
    description: Local API server
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: PSK token from ~/.quiet/{peer_id}/api_token
paths:
...
  /networks/{network_id}:
    get:
      summary: Get network details
      operationId: getNetwork
      x-module: network.get_for_peer
      parameters:
        - name: network_id
          in: path
          required: true
          schema:
            type: string

  # ============== Channels ==============
...
  /networks/join:
    post:
      summary: Join network via invite link
      operationId: joinNetwork
      x-module: user.join
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [invite_link, username]
              properties:
                invite_link:
                  type: string
                username:
                  type: string
                device_name:
                  type: string
                  default: "Device"
      responses:
        201:
          description: Joined network
          content:
            application/json:
              schema:
                type: object
                properties:
                  network_id:
                    type: string
                  user_id:
                    type: string
                  peer_shared_id:
                    type: string

  # ============== Files ==============
  /networks/{network_id}/files/{file_id}:
    get:
      summary: Download complete file
      operationId: downloadFile
      x-module: message_attachment.get_file_data
      parameters:
        - name: network_id
          in: path
          required: true
          schema:
            type: string
        - name: file_id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: File binary
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
  /networks/{network_id}/files/{file_id}/status:
    get:
      summary: Get file sync status
      operationId: getFileStatus
      x-module: message_attachment.get_file_download_progress
      parameters:
        - name: network_id
          in: path
          required: true
          schema:
            type: string
        - name: file_id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: File status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileStatus'
  /networks/{network_id}/files/{file_id}/data:
    get:
      summary: Read partial file data
      operationId: readFileData
      x-module: message_attachment.get_file_data (with offset/length)
      parameters:
        - name: network_id
          in: path
          required: true
          schema:
            type: string
        - name: file_id
          in: path
          required: true
          schema:
            type: string
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: length
          in: query
          schema:
            type: integer
      responses:
        200:
          description: Partial file binary
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
  /networks/{network_id}/files/{file_id}/sync:
    post:
      summary: Request file sync with priority
      operationId: requestFileSync
      x-module: sync_file (priority sync)
      parameters:
        - name: network_id
          in: path
          required: true
          schema:
            type: string
        - name: file_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                priority:
                  type: integer
                  minimum: 1
                  maximum: 10
                  default: 5

  # ============== Sync Status ==============
...
components:
  schemas:
...
    FileStatus:
      type: object
      properties:
        file_id:
          type: string
        status:
          type: string
          enum: [syncing, complete, failed]
        progress:
          type: number
          minimum: 0
          maximum: 1
        bytes_synced:
          type: integer
        bytes_total:
          type: integer
...